Moved create task into new command.
