+++ src/projects/index.js
@@ -1,24 +1,15 @@
-// Native
-const {join, relative} = require('path');
-
 // External
-const copyTemplateDir = require('copy-template-dir');
-const pify = require('pify');
 const pkgDir = require('pkg-dir');
 const readPkg = require('read-pkg');
 
 // Ours
-const ourPkg = require('../../package');
 const {packs, prequire, unpack} = require('../utils/async');
-const {hvy, ok} = require('../utils/color');
-const {log} = require('../utils/console');
-const {isRepo, createRepo, getCurrentLabel} = require('../utils/git');
+const {hvy} = require('../utils/color');
+const {isRepo, getCurrentLabel} = require('../utils/git');
 const {pretty} = require('../utils/misc');
-const {install} = require('../utils/npm');
-const {BUILD_DIR, CONFIG_FILE_NAME, DEFAULTS, KNOWN_TARGETS} = require('./constants');
+const {CONFIG_FILE_NAME, DEFAULTS, KNOWN_TARGETS} = require('./constants');
 
 // Wrapped
-const clone = packs(pify(copyTemplateDir));
 const getPkg = packs(readPkg);
 const getPkgDir = packs(pkgDir);
 
@@ -27,52 +18,10 @@ let root;
 let pkg;
 let config;
 
-const [MAJOR, MINOR] = ourPkg.version.split('.');
-const DEFAULT_TEMPLATE_VARS = {
-  auntyVersion: [MAJOR, MINOR, 'x'].join('.'),
-  authorName: 'ABC News',
-  buildDir: BUILD_DIR,
-  currentYear: (new Date()).getFullYear()
-};
-
 const MESSAGES = {
   NO_CONFIG: `This project has no ${hvy(CONFIG_FILE_NAME)} file or ${hvy('aunty')} property in its ${hvy('package.json')}.`,
-  NOT_PACKAGE: `This command can only be run inside a project with a ${hvy('package.json')} file.`,
-  creating: (type, dir, vars) => pretty`
-Creating a ${hvy(type)} project in:
-
-${hvy(dir)}
-
-...using template variables:
-
-${vars}
-  `,
-  created: (dir, file) => `  ${ok('create')} ${relative(dir, file)}`
+  NOT_PACKAGE: `This command can only be run inside a project with a ${hvy('package.json')} file.`
 };
-
-module.exports.create = packs(async config => {
-  const projectTypeConfig = require(`./${config.projectType}`);
-  const templateDir = join(__dirname, `../../templates/${config.projectType}`);
-  const targetDir = join(process.cwd(), config.directoryName);
-  const templateVars = Object.assign({}, DEFAULT_TEMPLATE_VARS, config.templateVars);
-
-  log(MESSAGES.creating(config.projectType, targetDir, templateVars));
-
-  const files = unpack(await clone(templateDir, targetDir, templateVars));
-
-  files.sort().forEach(file => log(MESSAGES.created(targetDir, file)));
-
-  log('Installing dependencies…');
-  await install(['--only=dev'], targetDir);
-  await install(['--save'].concat(projectTypeConfig.dependencies), targetDir);
-
-  if (await isRepo(targetDir)) {
-    return log('Git repo already exists');
-  }
-
-  log('Creating git repo…');
-  await createRepo(targetDir);
-});
 
 module.exports.getConfig = packs(async argv => {
   if (!root) {
